// src/api/model/user-view-dto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Vector
 * The Vector API
 * OpenAPI spec version: 1.0
 */
import type { UserViewDashboardDto } from "./user-view-dashboard-dto"
import type { UserViewTableDto } from "./user-view-table-dto"

export interface UserViewDto {
	table: UserViewTableDto
	dashboard: UserViewDashboardDto
}